// src/model/news.ts
export interface INews {
    featured:boolean;
            
            img: string;
            title: string;
            desc: string;
            category:string;
            id: string;
            createddate: string;
            publishdate: string;
            archivedate: string;
            customurl: string;
            author:  string;
            statusid: number;
            status: string;
            content:string;
            bannerscale: string;
            featuredscale: string;
            thumb1scale: string;
            thumb2scale: string;
            bannerx: string;
            bannery: string;
            featuredx: string;
            featuredy: string;
            thumb1x: string;
            thumb1y: string;
            thumb2x: string;
            thumb2y: string;
            privacy: any;
}